Introduced rotating seeds for session thread tracking
We use seeds to stitch together user journeys.  In a perfect world,
the (ip_hash, tag) pair would be unique to a visitor, even though IP
address may change be over time.  But it is quite reasonable to
imagine multiple visitors arriving from behind the same NAT address.
In such a case, we'd like to be able to knit together the pages each
individual visited.

Our strategy is to generate a new seed on each visit.  All outgoing
links, whether local (/D/...) or exit (/E/...) take a tag and then a
seed.  On each visit we record referrer and the URL requested.  When
stitching user journeys, we process (ip_hash, tag) pair threads and
then construct the forest of trees that have the property that child
nodes record as referrer the this_page_url of the parent.

Part of #10.